requirements_web.txt
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
transformers==4.35.0
ultralytics==8.0.200
matplotlib==3.8.2
numpy==1.26.2
requests==2.31.0
huggingface-hub==0.19.4
